cli/pyproject.toml
[tool.poetry]
name = "chip-certification-tool-cli"
version = "0.1.0"
description = "A CLI tool for accessing CHIP Test Harness"

authors = []

readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
httpx = ">=0.15.4,<0.19.0"
attrs = ">=20.1.0,<22.0.0"
python-dateutil = "^2.8.0"
websockets = "^10.0"
click = "^8.0.1"
pydantic = "^1.4"
types-click = "^7.1.5"
fastapi = "^0.68.1"
loguru = "^0.5.3"
aioconsole = "^0.3.2"
PyYAML = "^5.4.1"


[tool.poetry.dev-dependencies]
mypy = "^1.2.0"
black = "^22.3.0"
isort = "^5.12.0"
autoflake = "^2.1.1"
flake8 = "^6.0.0"
types-python-dateutil = "^2.8.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.black]
line-length = 120
exclude = '''
/(
  | .git
  | __pycache__
  | __init__.py
  | .mypy_cache
  | .pytest_cache
  | .venv
  | client_generator
)/
'''

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
